feat: a second pass on git.rs
